Updated to support new v1.8 deployment
We no longer generate config when the Topaz container is bootstrapped, rather the conf volume must be preconfigured. The bin/create-conf-volume.sh script will take care of this, with various parameters visible therein.

Also JVMs are now broken out into separate containers, as shown in the Docker compose scripts. Since we have moved to Java 11 in Topaz 1.8 the memory settings are now read from the container, hence the mem_limit entries in the compose file (and also why we had to downgrade to compose file version 2.x - limits are swarm only in 3.x). Previously this was set in the application.vmoptions file inside the conf volume.